changes to structure, everything went back to main.hs, to fix import issues
