+++ app/Main.hs
@@ -1,10 +1,9 @@
 module Main where
 
+import System.Random
 import Repo.Usuarios
 import Repo.Personas
-import Repo.Logic
-import Repo.PlayerxBot
-import Repo.PlayerxPlayer
+
 
 
 main :: IO ()
@@ -91,10 +90,278 @@ menuComoJogar = do
     opcaoEscolhidaCJ <- getLine
 
     if opcaoEscolhidaCJ == "1" then do
-        pegaEscolhaInicial
+        main
         else do
             putStr ("Opcao invalida! \n")
-            menuComoJogar --para que o programa não termine, e retorne ao mesmo menu.
+            menuComoJogar 
+
+---------------
+--PlayerxPlayer:
+----------------
+novoJogoPlayerXPlayer :: IO ()
+novoJogoPlayerXPlayer = do
+    cls
+    putStr "|================================================================|\n"
+    putStr "|==========================| NOVO JOGO |=========================|\n"
+    putStr "|================================================================|\n\n"
+
+    print "Digite o nome do jogador 1"
+    username1 <- getLine
+
+    print "Digite o nome do jogador 2"
+    username2 <- getLine
+
+    if username1 == username2 then do
+        putStr "ERRO: Nomes nao podem ser iguais. Pressione Enter para continuar...\n"
+        getLine
+        novoJogoPlayerXPlayer
+    else do
+        userField <- criaUsuarioSeNaoExistir username1
+        let user1 = parseUserToStringArray (head userField)
+
+        userField <- criaUsuarioSeNaoExistir username2
+        let user2 = parseUserToStringArray (head userField)
+
+        jogoPxP user1 user2
+
+        main
+
+
+jogoPxP :: [String] -> [String] -> IO()
+jogoPxP user1 user2 = do
+    personasField <- listPersonas
+    let personas = map parsePersonaToStringArray personasField
+
+    escolha <- escolhePersonagem user1 personas
+    let index = (read escolha :: Int) -1
+    let personaUser = personas!!index
+
+    escolha <- escolhePersonagem user2 personas
+    let index = (read escolha :: Int) -1
+    let personaUser2 = personas!!index
+
+    let possibilidadesUser = map parsePersonaToStringArray personasField
+    let possibilidadesUser2 = map parsePersonaToStringArray personasField
+
+    partidaPxP user1 user2 personaUser personaUser2 possibilidadesUser possibilidadesUser2
+
+    main
+
+
+partidaPxP :: [String] -> [String] -> [String] -> [String] -> [[String]] -> [[String]] -> IO()
+partidaPxP jogadorDaVez oponente personaJogadorDaVez personaOponente possibilidadesJogadorDaVez possibilidadesOponente = do
+    if userGanhou possibilidadesOponente then do
+        vitoria oponente jogadorDaVez personaJogadorDaVez
+    else do
+        cls
+        putStr "|===================================================================================|\n"
+        printTable possibilidadesJogadorDaVez
+        putStr "|===================================================================================|\n\n"
+
+        palpite <- pegarPalpiteUser jogadorDaVez
+        let newPossibilidades = verificarPalpite palpite personaOponente possibilidadesJogadorDaVez
+        if length newPossibilidades == length possibilidadesJogadorDaVez then do
+            putStr "OPCAO INVALIDA! Pressione Enter para tentar novamente...\n"
+            getLine
+            partida jogadorDaVez oponente personaJogadorDaVez personaOponente possibilidadesJogadorDaVez possibilidadesOponente
+        else
+            partida oponente jogadorDaVez personaOponente personaJogadorDaVez possibilidadesOponente newPossibilidades
+
+
+vitoria :: [String] -> [String] -> [String] -> IO()
+vitoria userGanhador oponente personaOponente = do
+    somaPontosUsuario (head userGanhador) 1
+
+    cls
+    putStr "|=======================================================================================|\n"
+    putStr "|======================================| VICTORY! |=====================================|\n"
+    putStr "|=======================================================================================|\n\n"
+    putStr ("PARABENS " ++ head userGanhador ++ "! VOCE GANHOU!\n")
+    putStr ("O personagem de " ++ head oponente ++ " era: " ++ prepareTableLine personaOponente ++ "\n")
+    putStr "Pressione Enter para voltar ao menu principal...\n"
+    getLine
+    putStr ""
+
+---------------
+--PlayerxBot:
+----------------
+novoJogoPlayerXBot :: IO ()
+novoJogoPlayerXBot = do
+    cls
+    putStr "|================================================================|\n"
+    putStr "|==========================| NOVO JOGO |=========================|\n"
+    putStr "|================================================================|\n\n"
+
+    print "Digite o nome do jogador "
+    username1 <- getLine
+
+    if username1 == "" then do
+        putStr "ERRO: É necessário digitar um nome de usuário, aperte ENTER para continuar \n"
+        getLine
+        novoJogoPlayerXBot
+    else do
+        userField <- criaUsuarioSeNaoExistir username1
+        let user1 = parseUserToStringArray (head userField)
+
+        jogoPxB user1 
+
+        main
+
+
+jogoPxB :: [String] -> IO()
+jogoPxB user1 = do
+    personasField <- listPersonas
+    let personas = map parsePersonaToStringArray personasField
+
+    escolha <- escolhePersonagem user1 personas
+    let index = (read escolha :: Int) -1
+    let personaUser = personas!!index
+
+    --Gera um numero randomico, que definirá o personagem do bot
+    indexBot <- randomIO :: IO Int
+    let index = indexBot `mod` (length personas)       
+    let personaBot = personas!!index
+
+    let possibilidadesUser = map parsePersonaToStringArray personasField
+    let possibilidadesBot = map parsePersonaToStringArray personasField
+
+    partidaPxB user1 personaUser personaBot possibilidadesUser possibilidadesBot
+
+    main
+
+
+partidaPxB :: [String] -> [String] -> [String] -> [[String]] -> [[String]] -> IO()
+partidaPxB jogador personaJogador personaBot possibilidadesJogador possibilidadesBot = do
+    if userGanhou possibilidadesBot then do
+        vitoriaBot personaBot
+    else if userGanhou possibilidadesJogador then do
+        vitoriaJogador jogador personaBot
+    else do
+        cls
+        putStr "|===================================================================================|\n"
+        printTable possibilidadesJogador
+        putStr "|===================================================================================|\n\n"
+
+        palpite <- pegarPalpiteUser jogador
+        let newPossibilidadesJogador = verificarPalpite palpite personaBot possibilidadesJogador
+        if length newPossibilidadesJogador == length possibilidadesJogador then do
+            putStr "OPCAO INVALIDA! Pressione Enter para tentar novamente...\n"
+            getLine
+            partidaPxB jogador personaJogador personaBot possibilidadesJogador possibilidadesBot
+        else
+            palpiteIa <- pegarPalpiteIA possibilidadesBot
+            let newPossibilidadesBot = verificarPalpite palpiteIa personaJogador possibilidadesBot
+            partidaPxB jogador personaJogador personaBot newPossibilidadesJogador newPossibilidadesBot
+
+vitoriaJogador :: [String] -> [String] -> IO()
+vitoriaJogador jogador personaBot = do
+    somaPontosUsuario (head jogador) 1
+
+    cls
+    putStr "|=======================================================================================|\n"
+    putStr "|======================================| VICTORY! |=====================================|\n"
+    putStr "|=======================================================================================|\n\n"
+    putStr ("Você venceu!\n")
+    putStr ("O personagem do bot era: " ++ prepareTableLine personaBot ++ "\n")
+    putStr "Pressione Enter para voltar ao menu principal...\n"
+    getLine
+    putStr ""
+
+
+vitoriaBot :: [String] -> IO()
+vitoriaBot personaBot = do
+    cls
+    putStr "|=======================================================================================|\n"
+    putStr "|======================================| DEFEAT! |=====================================|\n"
+    putStr "|=======================================================================================|\n\n"
+    putStr ("Você PERDEU!\n")
+    putStr ("O personagem do bot era: " ++ prepareTableLine personaBot ++ "\n")
+    putStr "Pressione Enter para voltar ao menu principal...\n"
+    getLine
+    putStr ""
+
+---------------
+--Logic:
+----------------
+escolhePersonagem :: [String] -> [[String]] -> IO String
+escolhePersonagem user personas = do
+    cls
+    putStr "|===================================================================================|\n"
+    putStr "|=============================| ESCOLHA SEU PERSONAGEM |============================|\n"
+    putStr "|===================================================================================|\n"
+    printTable personas
+    putStr "\n"
+    putStr ("Escolha um personagem, " ++ user!!0 ++ ":\n")
+
+    getLine
+
+
+userGanhou :: [[String]] -> Bool
+userGanhou listaPossibilidades = length listaPossibilidades <= 1
+
+
+pegarPalpiteUser :: [String] -> IO String
+pegarPalpiteUser user = do
+    putStr ("Qual caracteristica voce deseja escolher " ++ head user ++ "?\n")
+    putStr "[0] sexo | [1] cor do cabelo | [2] etnia | [3] cor dos olhos | [4] acessorio\n"
+    caracteristica <- getLine
+
+    if caracteristica == "0" then do
+        putStr "Qual o sexo?\n"
+        getLine
+    else if caracteristica == "1" then do
+        putStr "Qual a cor do cabelo?\n"
+        getLine
+    else if caracteristica == "2" then do
+        putStr "Qual a etnia?\n"
+        getLine
+    else if caracteristica == "3" then do
+        putStr "Qual a cor dos olhos?\n"
+        getLine
+    else if caracteristica == "4" then do
+        putStr "Qual o acessorio?\n"
+        getLine
+    else do
+        putStr "OPCAO INVALIDA!\n"
+        pegarPalpiteUser user
+
+
+verificarPalpite :: String -> [String] -> [[String]] -> [[String]]
+verificarPalpite palpite persona possibilidades = do
+    if palpite `elem` persona then
+        [personagem | personagem <- possibilidades, palpite `elem` personagem]
+    else
+        [personagem | personagem <- possibilidades, palpite `notElem` personagem]
+
+
+printTable :: [[String]] -> IO()
+printTable = mapM_ (putStr . prepareTableLine)
+
+
+prepareTableLine :: [String] -> String
+prepareTableLine [] = ""
+prepareTableLine [head] = head ++ "\n"
+prepareTableLine (head:users) = "   " ++ head ++ "  |  " ++ prepareTableLine users
+
+
+-------------------------
+--Algoritmo da IA abaixo:
+-------------------------
+renderRandom::[a] -> IO Int
+renderRandom array =
+   randomRIO (0, length array - 1)
+
+
+getElemByIndex :: [a] -> IO a
+getElemByIndex list = do
+    index <- renderRandom list
+    return $ list!!index
+
+
+pegarPalpiteIA :: [[String]] -> IO String
+pegarPalpiteIA possibilidades = do
+    p <- getElemByIndex possibilidades
+    getElemByIndex (tail (tail p))
 
 
 --Função para limpar o terminal.
