+++ whoami.cabal
@@ -9,7 +9,7 @@ extra-source-files:
 executable whoami
     main-is:          Main.hs
   
-    other-modules:    Repo.Usuarios, Repo.Personas, Repo.Logic, Repo.PlayerxPlayer, Repo.PlayerxBot
+    other-modules:    Repo.Usuarios, Repo.Personas
     build-depends:    base, sqlite-simple, random
     hs-source-dirs:   app
     default-language: Haskell2010
